refactor: rename reminder-utils into utils
