+++ src/commands/reminder/update.ts
@@ -2,7 +2,7 @@ import { SlashCommandSubcommandBuilder } from 'discord.js';
 import { Result } from 'oxide.ts';
 import { convertDateToEpoch } from '../../utils/date-utils';
 import type { CommandHandler, Subcommand } from '../builder';
-import { updateReminder } from './reminder-utils';
+import { updateReminder } from './utils';
 
 export const data = new SlashCommandSubcommandBuilder()
   .setName('update')
